pyproject.toml
[tool.poetry]
name = "el-amanecerv3"
version = "1.0.0"
description = "Enterprise AI System with MCP Orchestrator, RAG, and Consciousness"
authors = ["Balmaurin <admin@sheily.ai>"]
readme = "README.md"
license = "MIT"
packages = [
    { include = "packages" },
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.4.0"
pydantic-settings = "^2.0.3"
sqlalchemy = "^2.0.23"
asyncpg = "^0.29.0"
redis = "^5.0.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
aiofiles = "^23.2.1"
httpx = "^0.25.0"
websockets = "^12.0"

# AI/ML
torch = "^2.1.0"
transformers = "^4.35.0"
sentence-transformers = "^2.2.2"
llama-cpp-python = "^0.2.20"
openai = "^1.3.0"

# RAG
qdrant-client = "^1.7.0"
chromadb = "^0.4.18"
langchain = "^0.0.340"
langchain-community = "^0.0.1"

# Monitoring
prometheus-client = "^0.19.0"
sentry-sdk = {extras = ["fastapi"], version = "^1.38.0"}

# Utils
python-dotenv = "^1.0.0"
pyyaml = "^6.0.1"
jinja2 = "^3.1.2"
loguru = "^0.7.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
pytest-xdist = "^3.5.0"
httpx = "^0.25.0"

# Code Quality
black = "^23.11.0"
ruff = "^0.1.6"
mypy = "^1.7.0"
isort = "^5.12.0"
pre-commit = "^3.5.0"

# Type stubs
types-redis = "^4.6.0"
types-passlib = "^1.7.7"
types-python-jose = "^3.3.4"

# Documentation
mkdocs = "^1.5.3"
mkdocs-material = "^9.4.14"
mkdocstrings = {extras = ["python"], version = "^0.24.0"}

# Security
pip-audit = "^2.6.1"
bandit = "^1.7.5"

[tool.poetry.group.training.dependencies]
torch = "^2.1.0"
torchvision = "^0.16.0"
pytorch-lightning = "^2.1.0"
wandb = "^0.16.0"
tensorboard = "^2.15.0"
accelerate = "^0.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[project]
name = "enterprise-ai-testing"
version = "1.0.0"
description = "Enterprise AI Testing Suite"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pytest>=7.0.0",
    "numpy>=1.21.0",
    "psutil>=5.9.0",
    "typing-extensions>=4.0.0"
]

# =====================================
# PYTEST CONFIGURATION
# =====================================

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --disable-warnings"
asyncio_mode = "auto"
markers = [
    "enterprise: Enterprise-grade test validation",
    "security: Security and compliance tests",
    "performance: Performance and scalability tests",
    "blockchain: Blockchain and smart contract tests",
    "api: API functionality tests",
    "rag: RAG system tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

# =====================================
# COVERAGE CONFIGURATION
# =====================================

[tool.coverage.run]
source = ["backend", "sheily_core", "packages"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/migrations/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

# =====================================
# RUFF CONFIGURATION (Linter)
# =====================================

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "venv",
    "build",
    "dist",
    "migrations",
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py

# =====================================
# BLACK CONFIGURATION (Formatter)
# =====================================

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | venv
  | build
  | dist
  | migrations
)/
'''

# =====================================
# MYPY CONFIGURATION (Type Checker)
# =====================================

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true
strict_optional = true

[[tool.mypy.overrides]]
module = [
    "tests.*",
    "scripts.*",
]
ignore_errors = true

# =====================================
# ISORT CONFIGURATION (Import Sorter)
# =====================================

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_gitignore = true
known_first_party = ["apps", "packages", "backend", "sheily_core"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

